+++ src/tests/mocks/clinicalTrialsGov.handler.ts
@@ -5,7 +5,7 @@ import { responseJSON } from './information/clinicalTrialsGovInfo';
 const BASE_URL = 'https://clinicaltrials.gov';
 
 export const clinicalTrialsGovHandlers = [
-  rest.get(`${BASE_URL}/api/v2/studies?pageSize=2`, (request, response, context) => {
+  rest.get(`${BASE_URL}/api/v2/studies`, (request, response, context) => {
     return response(context.body(JSON.stringify(responseJSON)));
   }),
 ];
